chore: add developer function helpers, update validateApiKey middleware #627
